Refactor KVMVCpu::Init for x86_64vCPU
